backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
bcrypt
langgraph
langchain-openai
langchain-core
pydantic
